Reorganisation of board crate
