+++ engine/board/src/cache/mod.rs
@@ -2,9 +2,9 @@ pub use constraints::MoveConstraints;
 use myopic_core::enum_map::EnumMap;
 use myopic_core::BitBoard;
 
-use crate::private::cache::rays::RaySet;
 use crate::TerminalState;
 use crate::{Board, MoveComputeType};
+use rays::RaySet;
 
 mod constraints;
 mod control;
