+++ engine/board/src/lib.rs
@@ -1,23 +1,33 @@
 use std::cell::RefCell;
+use std::cmp::max;
 use std::fmt::Debug;
+use std::str::FromStr;
 
 use enum_map::EnumMap;
+use enumset::EnumSet;
 
-pub use moves::Move;
+use cache::CalculationCache;
+use history::History;
 use myopic_core::anyhow::Result;
 use myopic_core::enum_map::Enum;
 pub use myopic_core::*;
+use parse::patterns;
 pub use parse::uci::UciMove;
+use positions::Positions;
+use rights::Rights;
 
-use crate::enumset::EnumSet;
-use crate::private::cache::CalculationCache;
-use crate::private::history::History;
-use crate::private::positions::Positions;
-use crate::private::rights::Rights;
+use crate::anyhow::{anyhow, Error};
 
+mod cache;
+mod evolve;
+mod fen;
+mod history;
 mod moves;
 mod parse;
-mod private;
+mod positions;
+mod rights;
+#[cfg(test)]
+mod test;
 
 /// The start position of a chess game encoded in FEN format
 pub const START_FEN: &'static str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
@@ -25,6 +35,14 @@ pub const START_FEN: &'static str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR
 /// Return the start position of a standard game
 pub fn start() -> Board {
     START_FEN.parse().unwrap()
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub enum Move {
+    Standard { moving: Piece, from: Square, dest: Square, capture: Option<Piece> },
+    Enpassant { side: Side, from: Square, dest: Square, capture: Square },
+    Promotion { from: Square, dest: Square, promoted: Piece, capture: Option<Piece> },
+    Castle { corner: Corner },
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Ord, PartialOrd, Hash, Enum)]
@@ -152,7 +170,7 @@ impl Board {
 
     /// Returns the Zobrist hash of this position.
     pub fn hash(&self) -> u64 {
-        private::hash(&self.pieces, &self.rights, self.active, self.enpassant)
+        hash(&self.pieces, &self.rights, self.active, self.enpassant)
     }
 
     /// Return the active side in this position, i.e. the one whose turn it is.
@@ -206,7 +224,7 @@ impl Board {
     /// Return the specified components of the FEN encoding of this position
     /// in the given order with components separated by a space.
     pub fn to_fen_parts(&self, parts: &[FenPart]) -> String {
-        private::fen::to_fen_impl(self, parts)
+        fen::to_fen_impl(self, parts)
     }
 
     /// Return the complete FEN representation of this position.
@@ -244,11 +262,160 @@ impl PartialEq<Board> for Board {
     }
 }
 
+impl FromStr for Board {
+    type Err = Error;
+
+    fn from_str(fen: &str) -> Result<Self, Self::Err> {
+        if patterns::fen().is_match(&fen) {
+            let space_split: Vec<_> =
+                patterns::space().split(&fen).map(|s| s.trim().to_owned()).collect();
+            let active = space_split[1].parse::<Side>()?;
+            let curr_move = space_split[5].parse::<usize>()?;
+            Ok(Board {
+                pieces: space_split[0].parse()?,
+                active: space_split[1].parse()?,
+                rights: space_split[2].parse()?,
+                enpassant: parse::parse_option(space_split[3].as_str())?,
+                clock: space_split[4].parse::<usize>()?,
+                history: History::new(2 * (max(curr_move, 1) - 1) + (active as usize)),
+                cache: RefCell::new(CalculationCache::default()),
+            })
+        } else {
+            Err(anyhow!("Cannot parse FEN {}", fen))
+        }
+    }
+}
+
+pub(crate) fn hash(pos: &Positions, rights: &Rights, active: Side, ep: Option<Square>) -> u64 {
+    let mut result = pos.hash() ^ hash::side(active) ^ ep.map_or(0u64, |x| hash::enpassant(x));
+    rights.corners().for_each(|c| result ^= hash::zone(c));
+    result
+}
+
+#[cfg(test)]
+mod fen_test {
+    use crate::test::TestBoard;
+    use crate::Board;
+    use crate::{Side, Square::*};
+
+    use super::*;
+
+    fn test(expected: TestBoard, fen_string: String) -> Result<()> {
+        assert_eq!(Board::from(expected), fen_string.parse::<Board>()?);
+        Ok(())
+    }
+
+    #[test]
+    fn fen_to_board_case_1() -> Result<()> {
+        let fen = "r1br2k1/1pq1npb1/p2pp1pp/8/2PNP3/P1N5/1P1QBPPP/3R1RK1 w - - 3 19";
+        let board = TestBoard {
+            whites: vec![A3 | B2 | C4 | E4 | F2 | G2 | H2, C3 | D4, !!E2, D1 | F1, !!D2, !!G1],
+            blacks: vec![A6 | B7 | D6 | E6 | F7 | G6 | H6, !!E7, C8 | G7, A8 | D8, !!C7, !!G8],
+            castle_rights: Rights::empty(),
+            clock: 3,
+            active: Side::W,
+            enpassant: None,
+            history_count: 36,
+        };
+        test(board, String::from(fen))
+    }
+
+    #[test]
+    fn fen_to_board_case_2() -> Result<()> {
+        let fen = "rnb2rk1/ppp2ppp/4pq2/8/2PP4/5N2/PP3PPP/R2QKB1R w KQ - 2 9";
+        let board = TestBoard {
+            whites: vec![A2 | B2 | C4 | D4 | F2 | G2 | H2, !!F3, !!F1, A1 | H1, !!D1, !!E1],
+            blacks: vec![A7 | B7 | C7 | E6 | F7 | G7 | H7, !!B8, !!C8, A8 | F8, !!F6, !!G8],
+            castle_rights: Rights::side(Side::W),
+            clock: 2,
+            active: Side::W,
+            enpassant: None,
+            history_count: 16,
+        };
+        test(board, String::from(fen))
+    }
+
+    #[test]
+    fn fen_to_board_case_3() -> Result<()> {
+        let fen = "r1bqkbnr/ppp1pppp/n7/3pP3/8/8/PPPP1PPP/RNBQKBNR w KQkq d6 0 3";
+        let board = TestBoard {
+            whites: vec![
+                A2 | B2 | C2 | D2 | E5 | F2 | G2 | H2,
+                B1 | G1,
+                C1 | F1,
+                A1 | H1,
+                !!D1,
+                !!E1,
+            ],
+            blacks: vec![
+                A7 | B7 | C7 | D5 | E7 | F7 | G7 | H7,
+                A6 | G8,
+                C8 | F8,
+                A8 | H8,
+                !!D8,
+                !!E8,
+            ],
+            castle_rights: Rights::all(),
+            clock: 0,
+            active: Side::W,
+            enpassant: Some(D6),
+            history_count: 4,
+        };
+        test(board, String::from(fen))
+    }
+
+    #[test]
+    fn fen_to_board_case_4() -> Result<()> {
+        let fen = "r6k/p5pp/p1b2qnN/8/3Q4/2P1B3/PP4PP/R5K1 b - - 2 21";
+        let board = TestBoard {
+            whites: vec![A2 | B2 | C3 | G2 | H2, !!H6, !!E3, !!A1, !!D4, !!G1],
+            blacks: vec![A7 | A6 | G7 | H7, !!G6, !!C6, !!A8, !!F6, !!H8],
+            castle_rights: Rights::empty(),
+            clock: 2,
+            active: Side::B,
+            enpassant: None,
+            history_count: 41,
+        };
+        test(board, String::from(fen))
+    }
+}
+
+impl Reflectable for Board {
+    fn reflect(&self) -> Self {
+        let start_hash = Board::default().hash();
+        if self.history.historical_positions().next() == Some(start_hash) {
+            // If we played from the start position we can reflect properly
+            let mut reflected = Board::default();
+            reflected.active = Side::B;
+            for m in self.history.historical_moves() {
+                reflected.make(m).unwrap()
+            }
+            reflected
+        } else {
+            // Otherwise we started from some intermediate position and we cannot keep our history
+            let pieces = self.pieces.reflect();
+            let rights = self.rights.reflect();
+            let active = self.active.reflect();
+            let enpassant = self.enpassant.reflect();
+            //let hash = hash(&pieces, &rights, active, enpassant);
+            Board {
+                history: History::default(),
+                clock: self.clock,
+                pieces,
+                rights,
+                active,
+                enpassant,
+                cache: RefCell::new(CalculationCache::default()),
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod uci_conversion_test {
     use myopic_core::*;
 
-    use crate::moves::Move;
+    use super::Move;
 
     #[test]
     fn test_pawn_standard_conversion() {
