+++ engine/board/src/moves.rs
@@ -1,202 +1,33 @@
-use std::fmt::{Display, Formatter};
-use std::str::FromStr;
+use myopic_core::*;
 
-use myopic_core::{
-    anyhow::{Error, Result},
-    Corner, Line, Reflectable, Side,
-};
-
-use crate::{Piece, Square};
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum Move {
-    Standard { moving: Piece, from: Square, dest: Square, capture: Option<Piece> },
-    Enpassant { side: Side, from: Square, dest: Square, capture: Square },
-    Promotion { from: Square, dest: Square, promoted: Piece, capture: Option<Piece> },
-    Castle { corner: Corner },
-}
-
-impl Display for Move {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Move::Standard { moving, from, dest, capture, .. } => write!(
-                f,
-                "s{}{}{}{}",
-                moving,
-                from,
-                dest,
-                match capture {
-                    None => "-".to_string(),
-                    Some(p) => p.to_string(),
-                }
-            ),
-            Move::Promotion { from, dest, promoted, capture, .. } => write!(
-                f,
-                "p{}{}{}{}",
-                from,
-                dest,
-                promoted,
-                match capture {
-                    None => "-".to_string(),
-                    Some(p) => p.to_string(),
-                }
-            ),
-            Move::Enpassant { side, from, dest, capture, .. } => {
-                write!(f, "e{}{}{}{}", side, from, dest, capture)
-            }
-            Move::Castle { corner, .. } => write!(f, "c{}{:?}", corner.0, corner.1),
-        }
-    }
-}
+use crate::cache::MoveConstraints;
+use crate::MoveComputeType;
+use crate::Square::*;
+use crate::{Board, Move};
 
 impl Reflectable for Move {
     fn reflect(&self) -> Self {
         match self {
-            &Move::Standard { moving, dest, from, capture, .. } => Move::Standard {
+            &Move::Standard { moving, dest, from, capture } => Move::Standard {
                 moving: moving.reflect(),
                 dest: dest.reflect(),
                 from: from.reflect(),
                 capture: capture.reflect(),
             },
-            &Move::Promotion { from, dest, promoted, capture, .. } => Move::Promotion {
+            &Move::Promotion { from, dest, promoted, capture } => Move::Promotion {
                 from: from.reflect(),
                 dest: dest.reflect(),
                 promoted: promoted.reflect(),
                 capture: capture.reflect(),
             },
-            &Move::Enpassant { side, from, dest, capture, .. } => Move::Enpassant {
+            &Move::Enpassant { side, from, dest, capture } => Move::Enpassant {
                 side: side.reflect(),
                 from: from.reflect(),
                 dest: dest.reflect(),
                 capture: capture.reflect(),
             },
-            &Move::Castle { corner, .. } => Move::Castle { corner: corner.reflect() },
+            &Move::Castle { corner } => Move::Castle { corner: corner.reflect() },
         }
-    }
-}
-
-#[cfg(test)]
-impl Move {
-    pub fn from(s: &str) -> Result<Move> {
-        use myopic_core::anyhow::anyhow;
-        match s.chars().next() {
-            None => Err(anyhow!("Cannot parse move from empty string!")),
-            Some(t) => match t {
-                's' => Ok(Move::Standard {
-                    moving: slice(s, 1, 2).parse()?,
-                    from: slice(s, 3, 2).parse()?,
-                    dest: slice(s, 5, 2).parse()?,
-                    capture: parse_op(slice(s, 7, 2).as_str())?,
-                }),
-                'e' => Ok(Move::Enpassant {
-                    side: slice(s, 1, 1).parse()?,
-                    from: slice(s, 2, 2).parse()?,
-                    dest: slice(s, 4, 2).parse()?,
-                    capture: slice(s, 6, 2).parse()?,
-                }),
-                'p' => Ok(Move::Promotion {
-                    from: slice(s, 1, 2).parse()?,
-                    dest: slice(s, 3, 2).parse()?,
-                    promoted: slice(s, 5, 2).parse()?,
-                    capture: parse_op(slice(s, 7, 2).as_str())?,
-                }),
-                'c' => Ok(Move::Castle {
-                    corner: Corner(slice(s, 1, 1).parse()?, slice(s, 2, 1).parse()?),
-                }),
-                _ => Err(anyhow!("Cannot parse {} as a move", s)),
-            },
-        }
-    }
-}
-
-#[cfg(test)]
-fn slice(s: &str, skip: usize, take: usize) -> String {
-    s.chars().skip(skip).take(take).collect()
-}
-
-pub fn parse_op<F>(s: &str) -> Result<Option<F>>
-where
-    F: FromStr<Err = Error>,
-{
-    match s {
-        "-" => Ok(None),
-        _ => Ok(Some(FromStr::from_str(s)?)),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use myopic_core::{Class, Flank};
-
-    use crate::moves::Move;
-    use crate::{Piece, Square};
-
-    use super::*;
-
-    #[test]
-    fn standard() -> Result<()> {
-        assert_eq!(
-            Move::Standard {
-                moving: Piece(Side::W, Class::P),
-                from: Square::E2,
-                dest: Square::E4,
-                capture: None,
-            },
-            Move::from("swpe2e4-")?
-        );
-        assert_eq!(
-            Move::Standard {
-                moving: Piece(Side::B, Class::R),
-                from: Square::C4,
-                dest: Square::C2,
-                capture: Some(Piece(Side::W, Class::P)),
-            },
-            Move::from("sbrc4c2wp")?
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn promotion() -> Result<()> {
-        assert_eq!(
-            Move::Promotion {
-                from: Square::E7,
-                dest: Square::E8,
-                promoted: Piece(Side::W, Class::Q),
-                capture: None,
-            },
-            Move::from("pe7e8wq-")?
-        );
-        assert_eq!(
-            Move::Promotion {
-                from: Square::E7,
-                dest: Square::D8,
-                promoted: Piece(Side::W, Class::Q),
-                capture: Some(Piece(Side::B, Class::B)),
-            },
-            Move::from("pe7d8wqbb")?
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn enpassant() -> Result<()> {
-        assert_eq!(
-            Move::Enpassant {
-                side: Side::B,
-                from: Square::D4,
-                dest: Square::C3,
-                capture: Square::C4,
-            },
-            Move::from("ebd4c3c4")?
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn castle() -> Result<()> {
-        assert_eq!(Move::Castle { corner: Corner(Side::B, Flank::K) }, Move::from("cbk")?);
-        Ok(())
     }
 }
 
@@ -223,5 +54,205 @@ impl Move {
                 format!("{}{}{}", from, dest, class)
             }
         }
+    }
+}
+
+impl Board {
+    pub(crate) fn compute_moves_impl(&self, computation_type: MoveComputeType) -> Vec<Move> {
+        let constraints = self.move_constraints(computation_type);
+        let pawn_moves = self.compute_pawn_moves(&constraints);
+        let nbrqk_moves = self.compute_nbrqk_moves(&constraints);
+        let castle_moves = match computation_type {
+            MoveComputeType::All => self.compute_castle_moves(&constraints),
+            _ => Vec::with_capacity(0),
+        };
+        pawn_moves
+            .into_iter()
+            .chain(nbrqk_moves.into_iter())
+            .chain(castle_moves.into_iter())
+            .collect()
+    }
+
+    fn compute_nbrqk_moves(&self, constraints: &MoveConstraints) -> Vec<Move> {
+        let mut dest: Vec<Move> = Vec::with_capacity(40);
+        let (whites, blacks) = self.sides();
+        let unchecked_moves = |p: Piece, loc: Square| p.moves(loc, whites, blacks);
+        // Add standard moves for pieces which aren't pawns
+        for class in &[Class::N, Class::B, Class::R, Class::Q, Class::K] {
+            let piece = Piece(self.active, *class);
+            for location in self.pieces.locs(piece) {
+                let moves = unchecked_moves(piece, location) & constraints.get(location);
+                dest.extend(self.standards(piece, location, moves));
+            }
+        }
+        dest
+    }
+
+    fn standards(&self, moving: Piece, from: Square, dests: BitBoard) -> Vec<Move> {
+        dests
+            .iter()
+            .map(|dest| Move::Standard { moving, from, dest, capture: self.piece(dest) })
+            .collect()
+    }
+
+    fn promotions(&self, side: Side, from: Square, dests: BitBoard) -> Vec<Move> {
+        dests
+            .iter()
+            .flat_map(|dest| {
+                [Class::Q, Class::R, Class::B, Class::N].iter().map(move |&promoted| {
+                    Move::Promotion {
+                        from,
+                        dest,
+                        promoted: Piece(side, promoted),
+                        capture: self.piece(dest),
+                    }
+                })
+            })
+            .collect()
+    }
+
+    fn compute_pawn_moves(&self, constraints: &MoveConstraints) -> Vec<Move> {
+        let mut moves: Vec<Move> = Vec::with_capacity(20);
+        let (standard, enpassant, promotion) = self.separate_pawn_locs();
+        let (active_pawn, (whites, blacks)) = (Piece(self.active, Class::P), self.sides());
+        let compute_moves = |loc: Square| active_pawn.moves(loc, whites, blacks);
+
+        // Add moves for pawns which can only produce standard moves.
+        for location in standard | enpassant {
+            let targets = compute_moves(location) & constraints.get(location);
+            moves.extend(self.standards(active_pawn, location, targets));
+        }
+        let active = self.active;
+        for from in enpassant {
+            let dest = self.enpassant.unwrap();
+            let capture = dest.next(active.reflect().pawn_dir()).unwrap();
+            if constraints.get(from).contains(capture)
+                && self.enpassant_doesnt_discover_attack(from)
+            {
+                moves.push(Move::Enpassant { side: active, from, dest, capture });
+            }
+        }
+        for location in promotion {
+            let targets = compute_moves(location) & constraints.get(location);
+            moves.extend(self.promotions(self.active, location, targets));
+        }
+
+        moves
+    }
+
+    fn enpassant_doesnt_discover_attack(&self, enpassant_source: Square) -> bool {
+        let (active, passive) = (self.active, self.active.reflect());
+        let active_king = self.king(active);
+        let third_rank = passive.pawn_third_rank();
+        if !third_rank.contains(active_king) {
+            return true;
+        }
+        let (r, q) = (Piece(passive, Class::R), Piece(passive, Class::Q));
+        let potential_attackers = self.locs(&[r, q]) & third_rank;
+        let all_pieces = self.all_pieces();
+        for loc in potential_attackers {
+            let cord = BitBoard::cord(loc, active_king) & all_pieces;
+            if cord.size() == 4
+                && cord.contains(enpassant_source)
+                && cord.intersects(self.locs(&[Piece(passive, Class::P)]))
+            {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    fn separate_pawn_locs(&self) -> (BitBoard, BitBoard, BitBoard) {
+        let enpassant_source =
+            self.enpassant.map_or(BitBoard::EMPTY, |sq| enpassant_attack_squares(self.active, sq));
+        let promotion_rank = self.active.pawn_promoting_from_rank();
+        let pawn_locs = self.locs(&[Piece(self.active, Class::P)]);
+        (
+            pawn_locs - enpassant_source - promotion_rank,
+            pawn_locs & enpassant_source,
+            pawn_locs & promotion_rank,
+        )
+    }
+
+    fn compute_castle_moves(&self, constraints: &MoveConstraints) -> Vec<Move> {
+        let king_constraint = constraints.get(self.king(self.active));
+        let (whites, blacks) = self.sides();
+        let p1 = |c: Corner| king_constraint.subsumes(uncontrolled_req(c));
+        let p2 = |c: Corner| !(whites | blacks).intersects(unoccupied_req(c));
+        self.rights
+            .corners()
+            .filter(|&c| p1(c) && p2(c))
+            .filter(|&c| {
+                let Line(k_source, _) = Line::king_castling(c);
+                let Line(r_source, _) = Line::rook_castling(c);
+                self.piece(k_source).map(|p| p.1 == Class::K).unwrap_or(false)
+                    && self.piece(r_source).map(|p| p.1 == Class::R).unwrap_or(false)
+            })
+            .map(|corner| Move::Castle { corner })
+            .collect()
+    }
+}
+
+fn uncontrolled_req(Corner(side, flank): Corner) -> BitBoard {
+    match (side, flank) {
+        (Side::W, Flank::K) => E1 | F1 | G1,
+        (Side::W, Flank::Q) => E1 | D1 | C1,
+        (Side::B, Flank::K) => E8 | F8 | G8,
+        (Side::B, Flank::Q) => E8 | D8 | C8,
+    }
+}
+
+fn unoccupied_req(Corner(side, flank): Corner) -> BitBoard {
+    match (side, flank) {
+        (Side::W, Flank::K) => F1 | G1,
+        (Side::W, Flank::Q) => D1 | C1 | B1,
+        (Side::B, Flank::K) => F8 | G8,
+        (Side::B, Flank::Q) => D8 | C8 | B8,
+    }
+}
+
+fn enpassant_attack_squares(active: Side, enpassant_square: Square) -> BitBoard {
+    ADJACENTS[enpassant_square.file_index() as usize] & active.reflect().pawn_third_rank()
+}
+
+//pub fn compute_adjacent_files() -> Vec<BitBoard> {
+//    let mut dest: Vec<BitBoard> = Vec::new();
+//    for i in 0..8 {
+//        if i == 0 {
+//            dest.push(FILES[1]);
+//        } else if i == 7 {
+//            dest.push(FILES[6]);
+//        } else {
+//            dest.push(FILES[i + 1] | FILES[i - 1]);
+//        }
+//    }
+//    dest
+//}
+
+const ADJACENTS: [BitBoard; 8] = [
+    BitBoard(144680345676153346),
+    BitBoard(361700864190383365),
+    BitBoard(723401728380766730),
+    BitBoard(1446803456761533460),
+    BitBoard(2893606913523066920),
+    BitBoard(5787213827046133840),
+    BitBoard(11574427654092267680),
+    BitBoard(4629771061636907072),
+];
+
+#[cfg(test)]
+mod test_enpassant_source_squares {
+    use myopic_core::{Side, Square::*};
+
+    use super::enpassant_attack_squares;
+
+    #[test]
+    fn test() {
+        assert_eq!(H4 | F4, enpassant_attack_squares(Side::B, G3));
+        assert_eq!(!!G4, enpassant_attack_squares(Side::B, H3));
+        assert_eq!(!!B4, enpassant_attack_squares(Side::B, A3));
+        assert_eq!(H5 | F5, enpassant_attack_squares(Side::W, G6));
+        assert_eq!(!!G5, enpassant_attack_squares(Side::W, H6));
+        assert_eq!(!!B5, enpassant_attack_squares(Side::W, A6));
     }
 }
