+++ engine/board/src/test/moves/szukstra_tal.rs
