+++ engine/brain/src/search/mod.rs
@@ -16,9 +16,9 @@ use crate::{eval, Evaluator};
 mod movehints;
 mod movequality;
 pub mod negascout;
+mod quiescent;
 pub mod terminator;
 mod transpositions;
-mod quiescent;
 
 const DEPTH_UPPER_BOUND: usize = 10;
 const SHALLOW_EVAL_TRIGGER_DEPTH: usize = 2;
