+++ engine/brain/src/search/negascout.rs
@@ -7,11 +7,11 @@ use itertools::Itertools;
 use myopic_board::anyhow::{anyhow, Result};
 use myopic_board::{Move, MoveComputeType, TerminalState};
 
-use crate::search::{eval, quiescent};
 use crate::search::movehints::MoveOrderingHints;
 use crate::search::movequality::{BestMoveHeuristic, MaterialAndPositioningHeuristic};
 use crate::search::terminator::SearchTerminator;
 use crate::search::transpositions::{TranspositionTable, TreeNode};
+use crate::search::{eval, quiescent};
 use crate::Evaluator;
 
 /// Performs a negascout search without any iterative deepening,
